backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
numpy==2.2.4
scikit-learn==1.6.1
xgboost==2.1.4
pydantic==2.10.6
openai==1.68.2
python-multipart==0.0.20
python-dotenv==1.0.1
